adsf

+++ server/app/controllers/EventManager.java
@@ -273,7 +273,7 @@ public static Result rsvp_to_event()
 			try(Connection conn2 = DB.getConnection()) {
 				PreparedStatement stmt2 = conn2.prepareStatement("UPDATE `CampusFeed`.`Event_has_User` SET `rsvp` = '1' WHERE `event_has_user`.`event_id` = ? AND `event_has_user`.`user_id` = ?");
 				stmt2.setLong(2, user_id);
-				stmt2.setInt(1, event_id);
+				stmt2.setLong(1, event_id);
 				stmt2.executeUpdate();
 				
 				
@@ -299,7 +299,7 @@ public static Result rsvp_to_event()
 	try(Connection conn = DB.getConnection()) {
 		PreparedStatement stmt = conn.prepareStatement("INSERT INTO CampusFeed.Event_has_User (user_id,event_id,rsvp,is_admin) VALUES (?,?,1,0)");
 		stmt.setLong(1, user_id);
-		stmt.setInt(2, event_id);
+		stmt.setLong(2, event_id);
 		stmt.executeUpdate();
